Add forloop in the homapage with items in DB

+++ item/models.py
@@ -18,7 +18,7 @@ class Item(models.Model):
         Category, related_name='items', on_delete=models.CASCADE)
     name = models.CharField(max_length=255)
     description = models.TextField(blank=True, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=999, decimal_places=2)
     image = models.ImageField(upload_to='item_images', blank=True, null=True)
     is_sold = models.BooleanField(default=False)
     create_by = models.ForeignKey(
